Add get_resampling_function function
